app/requirements.txt
streamlit>=1.28.0
requests>=2.31.0
sseclient-py>=1.8.0
streamlit-agraph>=0.0.45
